# frontend/server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for Bitmoi service",
        "title": "Bitmoi API",
        "contact": {
            "name": "API Support",
            "email": "bitmoiigson@gmail.com"
        },
        "license": {
            "name": "CC BY-NC-SA 4.0",
            "url": "https://creativecommons.org/licenses/by-nc-sa/4.0/"
        },
        "version": "1.0"
    },
    "host": "api.bitmoi.co.kr",
    "basePath": "/",
    "paths": {
        "/bidToken": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "erc20"
                ],
                "summary": "광고 스팟에 MOI 토큰을 입찰합니다.",
                "parameters": [
                    {
                        "description": "입찰 금액과 광고 스팟",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.BidTokenRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.ScoreResponse"
                        }
                    }
                }
            }
        },
...
        "/highestBidder": {
            "get": {
                "tags": [
                    "erc20"
                ],
                "summary": "특정 광고 스팟에 가장 높은 가격을 제시한 입찰자를 반환합니다.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "광고 스팟",
                        "name": "location",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "최상위 입찰자",
                        "schema": {
                            "$ref": "#/definitions/api.HighestBidderResponse"
                        }
                    }
                }
            }
        },
...
        "/selectedBidder": {
            "get": {
                "tags": [
                    "erc20"
                ],
                "summary": "특정 광고 스팟의 입찰에 성공한 사용자를 불러옵니다.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "광고 스팟",
                        "name": "location",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "사용자와 입찰금액",
                        "schema": {
                            "$ref": "#/definitions/api.HighestBidderResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.BidTokenRequest": {
            "type": "object",
            "required": [
                "amount",
                "location"
            ],
            "properties": {
                "amount": {
                    "type": "integer",
                    "minimum": 1
                },
                "location": {
                    "type": "string",
                    "enum": [
                        "practice",
                        "rank",
                        "freetoken"
                    ]
                }
            }
        },
        "api.CandleData": {
            "type": "object",
            "properties": {
                "pdata": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.PriceData"
                    }
                },
                "vdata": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.VolumeData"
                    }
                }
            }
        },
...
        "api.HighestBidderResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "api.OrderResult": {
            "type": "object",
            "properties": {
                "commission": {
                    "type": "number"
                },
                "end_price": {
                    "type": "number"
                },
                "entry_price": {
                    "type": "number"
                },
                "entry_time": {
                    "type": "string"
                },
                "is_liquidated": {
                    "type": "boolean"
                },
                "leverage": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "out_time": {
                    "type": "integer"
                },
                "pnl": {
                    "type": "number"
                },
                "roe": {
                    "type": "number"
                },
                "stage": {
                    "type": "integer"
                }
            }
        },
        "api.PriceData": {
            "type": "object",
            "properties": {
                "close": {
                    "type": "number"
                },
                "high": {
                    "type": "number"
                },
                "low": {
                    "type": "number"
                },
                "open": {
                    "type": "number"
                },
                "time": {
                    "type": "integer"
                }
            }
        },
...
        "api.ScoreResponse": {
            "type": "object",
            "properties": {
                "origin_chart": {
                    "$ref": "#/definitions/api.CandleData"
                },
                "result_chart": {
                    "$ref": "#/definitions/api.CandleData"
                },
                "score": {
                    "$ref": "#/definitions/api.OrderResult"
                }
            }
        },
...
        "api.VolumeData": {
            "type": "object",
            "properties": {
                "color": {
                    "type": "string"
                },
                "time": {
                    "type": "integer"
                },
                "value": {
                    "type": "number"
                }
            }
        },
...
    }
}